Bugg fixad med javascript decimal, ändrade hours till decimal. Admin crud nästan klar saknas factory på services, Project Create är klar och Home lista projektet.

+++ Data/Entities/Summary.cs
@@ -17,7 +17,7 @@ namespace Data.Entities
         [Required]
         public int ProjectID { get; set; }
 
-        public int? TotalHours { get; set; }
+        public decimal? TotalHours { get; set; }
         public decimal? TotalPrice { get; set; }
         public string? Notes { get; set; }
